refactor(cli): rename workspace files to studio
- workspace.ts → studio.ts
- workspace-container.ts → studio-container.ts
- workspace.test.ts → studio.test.ts (x2)
- workspace-new.test.ts → studio-new.test.ts
- Export names: registerWorkspaceCommands → registerStudioCommands,
  registerWorkspaceContainerCommands → registerStudioContainerCommands
- Updated imports in cli.ts and test files
- Added development mode docs to README
