+++ packages/cli/src/cli.ts
@@ -18,8 +18,8 @@
 
 import { program } from 'commander';
 import chalk from 'chalk';
-import { registerWorkspaceCommands } from './commands/workspace.js';
-import { registerWorkspaceContainerCommands } from './commands/workspace-container.js';
+import { registerStudioCommands } from './commands/studio.js';
+import { registerStudioContainerCommands } from './commands/studio-container.js';
 import { registerAgentCommands } from './commands/agent.js';
 import { registerSessionCommands } from './commands/session.js';
 import { registerConfigCommands } from './commands/mcp.js';
@@ -164,8 +164,8 @@ program
   });
 
 // Register subcommand groups
-registerWorkspaceCommands(program);
-registerWorkspaceContainerCommands(program);
+registerStudioCommands(program);
+registerStudioContainerCommands(program);
 registerAgentCommands(program);
 registerSessionCommands(program);
 registerConfigCommands(program);
