+++ packages/cli/src/commands/studio-new.test.ts
