+++ packages/cli/src/studio.test.ts
